requirements.txt
# Telegram Bot Framework
# python-telegram-bot==20.7

# # Google Generative AI (Gemini)
# google-generativeai==0.3.2

# # Supabase Client
# supabase==2.3.4

# # Environment Variables
# python-dotenv==1.0.0

# # Additional utilities
# httpx==0.26.0

python-telegram-bot
google-generativeai
supabase
python-dotenv
# httpx se instalará solo, no hace falta ponerlo aquí
